Ignore cancelled events for events that can be cancelled.
The main intention of ignoring cancelling events is the new
feature that allows Vanish to automatically open chests
while you are vanished. This has repercussions if for
example the server is using LWC. If the player really
CANNOT access this chest, the event WILL be cancelled
but Vanish will still let them inside this chest they should
not be able to access.