Fix bug/conflict where other plugins such as OpenInv would cause shulkers to glitch out when opened in vanish.
The bug would result in the shulker appearing to be closed on the users screen, however when walking on it, it would act as if it were open and it would be appear open for other players as well. This shulker would not close until unloaded or broken.

Fixed by adding filtering out shulkers from regular onPlayerInteract and handling with a LOW event priority as to not conflict.